requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

langchain==0.2.1
langchain-core==0.2.5
langgraph==0.0.48

google-generativeai==0.5.4

chromadb==0.4.24

pytesseract==0.3.10
Pillow==10.2.0

paddleocr==2.7.0.3
paddlepaddle==2.6.1

opencv-python==4.9.0.80
yt-dlp==2024.4.9

requests==2.31.0
python-dotenv==1.0.1

redis==5.0.3

numpy==1.26.4
pandas==2.2.2

pydantic==2.7.1
